Rectifying event consumption in`on_touch_down`
This commit tries to rectify the assumption that a disabled widget having a touch collision should consume the event. This shall only be true if the widget is visible and disabled and not when the widget is disabled and not visible at all. This helps in a lot of scenarios where due to design constraints and animations, one widget might come on top of another while being hidden completely. As Kivy doesn't have equivalent feature of `pointer-events`, so we shouldn't consume the event if the widget is hidden.